refactor(view): Clear commentaries and useless code
