+++ src/view/components/block_text.rs
