+++ src/view/components/views/response/response_content_view.rs
@@ -1,5 +1,5 @@
 use crate::view::components::Component;
-use crate::view::components::TabBlockedText::TabBlockText;
+use crate::view::components::tab_blocked_text::TabBlockText;
 use crate::view::renderer::tui_rs::BackendTuiRs;
 use tui::layout::Rect;
