+++ src/view/components/views/tabs_request/mod.rs
@@ -1,5 +1,5 @@
 use crate::base::states::names::StatesNames;
-use crate::view::components::TabList::Tabslist;
+use crate::view::components::tab_list::Tabslist;
 use crate::view::renderer::tui_rs::BackendTuiRs;
 use crate::{base::stores::MainStore, view::components::Component};
 use tui::layout::Rect;
